fix logTCPPorts command, fix execa error handling

+++ packages/scripts/src/helpers/scripts.ts
@@ -756,7 +756,11 @@ export async function k8sStartService(
         const tempDir = fs.mkdtempSync(path.join(os.tmpdir(), 'tempYaml'));
         fs.writeFileSync(path.join(tempDir, `${serviceName}Deployment.yaml`), updatedYaml);
         const subprocess = await execaCommand(`kubectl create -n services-dev1 -f ${path.join(tempDir, `${serviceName}Deployment.yaml`)}`);
-        logger.debug(subprocess.stdout);
+        const { stdout, stderr } = subprocess;
+        if (stderr) {
+            throw new Error(stderr);
+        }
+        logger.debug(stdout);
         fs.rmSync(tempDir, { recursive: true, force: true });
     } catch (err) {
         logger.error(`The service ${serviceName} could not be started: `, err);
@@ -811,50 +815,76 @@ function waitForKafkaRunning(name: string, timeoutMs = 12000): Promise<void> {
 }
 
 export async function setAlias(tsPort: string) {
-    let subprocess = await execaCommand('earl aliases remove k8s-e2e 2> /dev/null || true', { shell: true });
-    logger.debug(subprocess.stdout);
-    subprocess = await execaCommand(`earl aliases add k8s-e2e http://${config.HOST_IP}:${tsPort}`);
-    logger.debug(subprocess.stdout);
+    try {
+        const subprocess1 = await execaCommand('earl aliases remove k8s-e2e 2> /dev/null || true', { shell: true });
+        logger.debug(subprocess1.stdout);
+        if (subprocess1.stderr) {
+            throw new Error(subprocess1.stderr);
+        }
+
+        const subprocess2 = await execaCommand(`earl aliases add k8s-e2e http://${config.HOST_IP}:${tsPort}`);
+        logger.debug(subprocess2.stdout);
+        if (subprocess2.stderr) {
+            throw new Error(subprocess2.stderr);
+        }
+    } catch (err) {
+        throw new Error(`Failed to set alias: ${err}`);
+    }
 }
 
 export async function showState(tsPort: string) {
-    const subprocess = await execaCommand('kubectl get deployments,po,svc --all-namespaces --show-labels -o wide');
-    logger.debug(subprocess.stdout);
-    logger.debug(await showESIndices());
-    logger.debug(await showAssets(tsPort));
+    try {
+        const subprocess = await execaCommand('kubectl get deployments,po,svc --all-namespaces --show-labels -o wide');
+        const { stdout, stderr } = subprocess;
+
+        logger.debug(stdout);
+        logger.debug(await showESIndices());
+        logger.debug(await showAssets(tsPort));
+        if (stderr) {
+            throw new Error(stderr);
+        }
+    } catch (err) {
+        signale.error(`Failed to get k8s resources: ${err}`);
+    }
 }
 
 async function showESIndices() {
-    const subprocess = await execaCommand(`curl -k ${config.SEARCH_TEST_HOST}/_cat/indices?v`);
-    return subprocess.stdout;
+    try {
+        const subprocess = await execaCommand(`curl -k ${config.SEARCH_TEST_HOST}/_cat/indices?v`);
+        const { stdout, stderr } = subprocess;
+        if (stderr) {
+            throw new Error(stderr);
+        }
+        return stdout;
+    } catch (err) {
+        signale.error(`Failed to retrieve indices: ${err}`);
+    }
 }
 
 async function showAssets(tsPort: string) {
     try {
         const subprocess = await execaCommand(`curl ${config.HOST_IP}:${tsPort}/v1/assets`);
-        return subprocess.stdout;
+        const { stdout, stderr } = subprocess;
+        if (stderr) {
+            throw new Error(stderr);
+        }
+        return stdout;
     } catch (err) {
-        return err;
+        signale.error(`Failed to curl assets: ${err}`);
     }
 }
 
 export async function logTCPPorts() {
     try {
-        let command: string;
-        let args: string[];
-
-        if (process.platform === 'darwin') {
-            command = 'netstat';
-            args = ['-an', '-f', 'inet', '-p', 'tcp'];
-        } else {
-            command = 'ss';
-            args = ['-tan4'];
+        const command = 'netstat -an | grep \'^tcp\' | awk \'{print $4}\' | tr ".:" " " | awk \'{print $NF}\' | sort -n | uniq | tr "\n" " "';
+        const subprocess = await execaCommand(command, { shell: true, reject: false });
+        const { stdout, stderr } = subprocess;
+        if (stderr) {
+            throw new Error(stderr);
         }
-
-        const { stdout } = await execa(command, args, { shell: true, reject: false });
-        signale.info('TCP Ports:\n', stdout);
+        signale.info(`TCP Ports currently in use:\n ${stdout}`);
     } catch (err) {
-        signale.error('Execa command failed trying to log ports: ', err);
+        signale.error(`Execa command failed trying to log ports: ${err}`);
     }
 }
 
@@ -862,7 +892,11 @@ export async function deletePersistentVolumeClaim(searchHost: string) {
     try {
         const label = searchHost.includes('opensearch') ? `app.kubernetes.io/instance=${searchHost}` : `app=${searchHost}-master`;
         const subprocess = await execaCommand(`kubectl delete -n services-dev1 pvc -l ${label}`);
-        logger.debug(`kubectl delete pvc: ${subprocess.stdout}`);
+        const { stdout, stderr } = subprocess;
+        if (stderr) {
+            throw new Error(stderr);
+        }
+        logger.debug(`kubectl delete pvc: ${stdout}`);
     } catch (err) {
         throw new TSError(`Failed to delete persistent volume claim:\n${err}`);
     }
@@ -877,7 +911,11 @@ export async function helmfileDestroy(selector: string) {
 
     try {
         const subprocess = await execaCommand(`helmfile destroy -f ${helmfilePath} --selector app=${selector}`);
-        logger.debug(`helmfile delete:\n${subprocess.stdout}`);
+        const { stdout, stderr } = subprocess;
+        if (stderr) {
+            throw new Error(stderr);
+        }
+        logger.debug(`helmfile destroy:\n${stdout}`);
     } catch (err) {
         logger.info(err);
     }
@@ -894,6 +932,9 @@ export async function helmfileCommand(command: string, clusteringType: 'kubernet
     let subprocess;
     try {
         subprocess = await execaCommand(`helmfile --state-values-file ${valuesPath} ${command} -f ${helmfilePath}`);
+        if (subprocess.stderr) {
+            throw new Error(subprocess.stderr);
+        }
     } catch (err) {
         throw new TSError(`Helmfile ${command} command failed:\n${err}`);
     } finally {
@@ -919,6 +960,9 @@ export async function determineSearchHost() {
     const filtered = serviceList.filter((svc: ServiceObj) => possible.includes(svc.name));
     if (filtered.length > 1) {
         throw new TSError('Multiple Possible Search Hosts Detected. Cannot reset store.');
+    }
+    if (filtered.length === 0) {
+        throw new TSError('No Search Host Detected. Cannot reset store.');
     }
     return filtered[0].name;
 }
